Fix locale variable resolution order
- LC_MESSAGES have precedence over LANG

- LANGAUGE is used as the last option

- Put output of `locale(1)` to the same rules of environment
  variable parsing

`gettext` uses LANGUAGES as the first option, but **only if**
the locale is set properly, this implementation only uses
LANGUAGE if no other variable is set.

Also, this patch adds a note to the README so users can notice
that the returned locale in POSIX systems is useful for
translations only.